+++ HISTORY.rst
@@ -10,7 +10,7 @@ Release History
 
 - Notebook metadata is filtered - only the most common metadata are stored in the text representation (#105)
 - New config option ``additional_metadata_on_text_files`` on the content manager. Defaults to ``True``. Change its value to ``False`` to avoid creating a YAML header or cell metadata if there was none initially (#110)
-- Language magic arguments are preserved (#111)
+- Language magic arguments are preserved in R Markdown, and also supported in ``light`` and ``percent`` scripts (#111)
 - First markdown cell exported as a docstring when using the Sphinx format (#107)
 
 0.8.3 (2018-10-19)